event: add encode and decode method for dml event (#397)
* event: remove useless codes


* event: rename some event and add version


* event: add decode and encode function


* event: add unit test for encode and decode dml event


* event: add unit test for encode and decode dml event 1


* fix error


---------

+++ pkg/common/event/resolved_ts_event.go
@@ -9,7 +9,8 @@ import (
 )
 
 type BatchResolvedEvent struct {
-	Version int
+	// Version is the version of the BatchResolvedEvent struct.
+	Version byte
 	Events  []ResolvedEvent
 }
 
@@ -71,7 +72,8 @@ func (b *BatchResolvedEvent) GetSize() int64 {
 
 // ResolvedEvent represents a resolvedTs event of a dispatcher.
 type ResolvedEvent struct {
-	Version      int
+	// Version is the version of the ResolvedEvent struct.
+	Version      byte
 	DispatcherID common.DispatcherID
 	ResolvedTs   common.Ts
 }
